test(news,scanner,wsh): builders + body-verifying test migration
PR 6 of the test-builder rollout.

- Add request builders in src/testdata/builders/{news,scanner,wsh}.rs;
  contract_news/broad_tape_news reuse MarketDataRequestBuilder.
- Migrate sync/async tests to assert_request<B>(builder); move inline
  mod tests blocks to sibling _tests.rs files.
- Drop self-loop encoder tests in {news,scanner,wsh}/common/encoders.rs
  and the inline wsh/mod.rs::common_tests encoder asserts.
- Remove dead MessageBus/AsyncMessageBus::request_messages trait
  methods (only consumers were the migrated sync tests, which now
  call MessageBusStub::request_messages directly).
