+++ src/wsh/async.rs
@@ -80,339 +80,5 @@ impl Client {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::common::test_utils::helpers::assert_proto_msg_id;
-    use crate::messages::OutgoingMessages;
-    use crate::stubs::MessageBusStub;
-    use crate::wsh::common::test_data::{self};
-    use std::sync::{Arc, RwLock};
-
-    #[tokio::test]
-    async fn test_wsh_metadata_table() {
-        use crate::wsh::common::test_tables::{wsh_metadata_test_cases, ApiExpectedResult};
-
-        for test_case in wsh_metadata_test_cases() {
-            let message_bus = Arc::new(MessageBusStub {
-                request_messages: RwLock::new(vec![]),
-                response_messages: test_case.response_messages,
-            });
-
-            let client = Client::stubbed(message_bus, test_case.server_version);
-            let result = client.wsh_metadata().await;
-
-            match test_case.expected_result {
-                ApiExpectedResult::Success { json } => {
-                    assert!(result.is_ok(), "Test '{}' failed: {:?}", test_case.name, result.err());
-                    assert_eq!(result.unwrap().data_json, json, "Test '{}' json mismatch", test_case.name);
-                }
-                ApiExpectedResult::ServerVersionError => {
-                    assert!(result.is_err(), "Test '{}' should have failed", test_case.name);
-                    assert!(
-                        matches!(result.unwrap_err(), Error::ServerVersion(_, _, _)),
-                        "Test '{}' wrong error type",
-                        test_case.name
-                    );
-                }
-            }
-        }
-    }
-
-    #[tokio::test]
-    async fn test_wsh_event_data_by_contract_table() {
-        use crate::wsh::common::test_tables::{event_data_by_contract_test_cases, ApiExpectedResult};
-
-        for test_case in event_data_by_contract_test_cases() {
-            let message_bus = Arc::new(MessageBusStub {
-                request_messages: RwLock::new(vec![]),
-                response_messages: test_case.response_messages,
-            });
-
-            let client = Client::stubbed(message_bus, test_case.server_version);
-            let result = client
-                .wsh_event_data_by_contract(
-                    test_case.contract_id,
-                    test_case.start_date,
-                    test_case.end_date,
-                    test_case.limit,
-                    test_case.auto_fill,
-                )
-                .await;
-
-            match test_case.expected_result {
-                ApiExpectedResult::Success { json } => {
-                    assert!(result.is_ok(), "Test '{}' failed: {:?}", test_case.name, result.err());
-                    assert_eq!(result.unwrap().data_json, json, "Test '{}' json mismatch", test_case.name);
-                }
-                ApiExpectedResult::ServerVersionError => {
-                    assert!(result.is_err(), "Test '{}' should have failed", test_case.name);
-                    assert!(
-                        matches!(result.unwrap_err(), Error::ServerVersion(_, _, _)),
-                        "Test '{}' wrong error type",
-                        test_case.name
-                    );
-                }
-            }
-        }
-    }
-
-    #[tokio::test]
-    async fn test_wsh_event_data_by_filter_subscription_table() {
-        use crate::wsh::common::test_tables::subscription_test_cases;
-
-        for test_case in subscription_test_cases() {
-            let message_bus = Arc::new(MessageBusStub {
-                request_messages: RwLock::new(vec![]),
-                response_messages: test_case.response_messages,
-            });
-
-            let client = Client::stubbed(message_bus, crate::server_versions::WSH_EVENT_DATA_FILTERS_DATE);
-            let mut subscription = client
-                .wsh_event_data_by_filter(test_case.filter, test_case.limit, test_case.auto_fill)
-                .await
-                .unwrap_or_else(|_| panic!("Test '{}' failed to create subscription", test_case.name));
-
-            let mut received_events = vec![];
-            while let Some(result) = subscription.next().await {
-                match result {
-                    Ok(event) => received_events.push(event.data_json),
-                    Err(e) => panic!("Test '{}' unexpected error: {:?}", test_case.name, e),
-                }
-            }
-
-            assert_eq!(
-                received_events.len(),
-                test_case.expected_events.len(),
-                "Test '{}' event count mismatch",
-                test_case.name
-            );
-
-            for (i, (received, expected)) in received_events.iter().zip(test_case.expected_events.iter()).enumerate() {
-                assert_eq!(received, expected, "Test '{}' event {} mismatch", test_case.name, i);
-            }
-        }
-    }
-
-    #[tokio::test]
-    async fn test_wsh_metadata_decode_table() {
-        use crate::messages::ResponseMessage;
-        use crate::subscriptions::{DecoderContext, StreamDecoder};
-        use crate::wsh::common::test_tables::WSH_METADATA_DECODE_TESTS;
-
-        for test_case in WSH_METADATA_DECODE_TESTS {
-            let mut message = ResponseMessage::from(test_case.message);
-            let result = WshMetadata::decode(&DecoderContext::default(), &mut message);
-
-            if test_case.should_error {
-                assert!(result.is_err(), "Test '{}' should have failed", test_case.name);
-                match test_case.error_type {
-                    Some("UnexpectedResponse") => assert!(matches!(result.unwrap_err(), Error::UnexpectedResponse(_))),
-                    _ => panic!("Unknown error type for test '{}'", test_case.name),
-                }
-            } else {
-                assert!(result.is_ok(), "Test '{}' failed: {:?}", test_case.name, result.err());
-                assert_eq!(
-                    result.unwrap().data_json,
-                    test_case.expected_json,
-                    "Test '{}' json mismatch",
-                    test_case.name
-                );
-            }
-        }
-    }
-
-    #[tokio::test]
-    async fn test_wsh_event_data_decode_table() {
-        use crate::messages::ResponseMessage;
-        use crate::subscriptions::{DecoderContext, StreamDecoder};
-        use crate::wsh::common::test_tables::WSH_EVENT_DATA_DECODE_TESTS;
-
-        for test_case in WSH_EVENT_DATA_DECODE_TESTS {
-            let mut message = ResponseMessage::from(test_case.message);
-            let result = WshEventData::decode(&DecoderContext::default(), &mut message);
-
-            if test_case.should_error {
-                assert!(result.is_err(), "Test '{}' should have failed", test_case.name);
-                match test_case.error_type {
-                    Some("Message") => assert!(matches!(result.unwrap_err(), Error::Message(_, _))),
-                    Some("UnexpectedResponse") => assert!(matches!(result.unwrap_err(), Error::UnexpectedResponse(_))),
-                    _ => panic!("Unknown error type for test '{}'", test_case.name),
-                }
-            } else {
-                assert!(result.is_ok(), "Test '{}' failed: {:?}", test_case.name, result.err());
-                assert_eq!(
-                    result.unwrap().data_json,
-                    test_case.expected_json,
-                    "Test '{}' json mismatch",
-                    test_case.name
-                );
-            }
-        }
-    }
-
-    #[tokio::test]
-    async fn test_wsh_event_data_by_filter_integration_table() {
-        use crate::wsh::common::test_tables::{event_data_by_filter_integration_test_cases, IntegrationExpectedResult};
-
-        for test_case in event_data_by_filter_integration_test_cases() {
-            let stub = Arc::new(MessageBusStub {
-                request_messages: RwLock::new(vec![]),
-                response_messages: test_case.response_messages,
-            });
-            let message_bus = stub.clone();
-
-            let client = Client::stubbed(message_bus, test_case.server_version);
-
-            let result = match test_case.name {
-                "successful filter request with autofill" => {
-                    let filter = "filter=value";
-                    let result = client
-                        .wsh_event_data_by_filter(
-                            filter,
-                            Some(100),
-                            Some(AutoFill {
-                                competitors: true,
-                                portfolio: false,
-                                watchlist: true,
-                            }),
-                        )
-                        .await;
-
-                    if result.is_ok() {
-                        let request_messages = stub.request_messages();
-                        assert_eq!(request_messages.len(), 1);
-                        assert_proto_msg_id(&request_messages[0], OutgoingMessages::RequestWshEventData);
-                    }
-                    result
-                }
-                "successful filter request without autofill" => {
-                    let filter = "filter=value";
-                    let result = client.wsh_event_data_by_filter(filter, None, None).await;
-
-                    if result.is_ok() {
-                        let request_messages = stub.request_messages();
-                        assert_eq!(request_messages.len(), 1);
-                        assert_proto_msg_id(&request_messages[0], OutgoingMessages::RequestWshEventData);
-                    }
-                    result
-                }
-                _ => client.wsh_event_data_by_filter("filter", None, None).await,
-            };
-
-            match test_case.expected_result {
-                IntegrationExpectedResult::Success => {
-                    assert!(result.is_ok(), "Test '{}' failed: {:?}", test_case.name, result.err());
-                }
-                IntegrationExpectedResult::ServerVersionError => {
-                    assert!(result.is_err(), "Test '{}' should have failed", test_case.name);
-                    if let Err(error) = result {
-                        assert!(
-                            matches!(error, Error::ServerVersion(_, _, _)),
-                            "Test '{}' wrong error type",
-                            test_case.name
-                        );
-                    }
-                }
-            }
-        }
-    }
-
-    #[tokio::test]
-    async fn test_server_version_validation_table() {
-        use crate::wsh::common::test_tables::server_version_test_cases;
-
-        for test_case in server_version_test_cases() {
-            let message_bus = Arc::new(MessageBusStub {
-                request_messages: RwLock::new(vec![]),
-                response_messages: vec![],
-            });
-
-            let client = Client::stubbed(message_bus, test_case.server_version);
-
-            let result = if let Some(contract_id) = test_case.contract_id {
-                client
-                    .wsh_event_data_by_contract(
-                        contract_id,
-                        test_case.start_date,
-                        test_case.end_date,
-                        test_case.limit,
-                        test_case.auto_fill,
-                    )
-                    .await
-            } else {
-                client
-                    .wsh_event_data_by_filter("filter", test_case.limit, test_case.auto_fill)
-                    .await
-                    .map(|_| WshEventData { data_json: "".to_string() })
-            };
-
-            if test_case.expected_error {
-                assert!(result.is_err(), "Test '{}' should have failed", test_case.name);
-                if let Err(error) = result {
-                    assert!(
-                        matches!(error, Error::ServerVersion(_, _, _)),
-                        "Test '{}' wrong error type",
-                        test_case.name
-                    );
-                }
-            } else {
-                assert!(result.is_ok(), "Test '{}' failed: {:?}", test_case.name, result.err());
-            }
-        }
-    }
-
-    #[tokio::test]
-    async fn test_subscription_integration_table() {
-        use crate::wsh::common::test_tables::subscription_integration_test_cases;
-
-        for test_case in subscription_integration_test_cases() {
-            let message_bus = Arc::new(MessageBusStub {
-                request_messages: RwLock::new(vec![]),
-                response_messages: test_case.response_messages,
-            });
-
-            let client = Client::stubbed(message_bus, test_case.server_version);
-            let result = client.wsh_event_data_by_filter("test_filter", None, None).await;
-
-            assert!(result.is_ok(), "Test '{}' failed to create subscription", test_case.name);
-            let mut subscription = result.unwrap();
-
-            // Collect all events
-            let mut events = vec![];
-            while let Some(event_result) = subscription.next().await {
-                match event_result {
-                    Ok(event) => events.push(event.data_json),
-                    Err(e) => panic!("Test '{}' unexpected error: {e:?}", test_case.name),
-                }
-            }
-
-            assert_eq!(
-                events.len(),
-                test_case.expected_events.len(),
-                "Test '{}' event count mismatch",
-                test_case.name
-            );
-
-            for (i, (received, expected)) in events.iter().zip(test_case.expected_events.iter()).enumerate() {
-                assert_eq!(received, expected, "Test '{}' event {} mismatch", test_case.name, i);
-            }
-        }
-    }
-
-    #[tokio::test]
-    async fn test_data_stream_cancel_message() {
-        use crate::subscriptions::StreamDecoder;
-
-        let request_id = test_data::REQUEST_ID_METADATA;
-
-        // Test WshMetadata cancel
-        let cancel_msg = WshMetadata::cancel_message(0, Some(request_id), None);
-        assert!(cancel_msg.is_ok());
-        assert_proto_msg_id(&cancel_msg.unwrap(), OutgoingMessages::CancelWshMetaData);
-
-        // Test WshEventData cancel
-        let cancel_msg = WshEventData::cancel_message(0, Some(request_id), None);
-        assert!(cancel_msg.is_ok());
-        assert_proto_msg_id(&cancel_msg.unwrap(), OutgoingMessages::CancelWshEventData);
-    }
-}
+#[path = "async_tests.rs"]
+mod tests;
