+++ src/wsh/mod.rs
@@ -11,8 +11,6 @@ use serde::{Deserialize, Serialize};
 mod common;
 
 // Re-export common functionality
-#[cfg(test)]
-use common::decoders;
 use common::encoders;
 
 // Feature-specific implementations
@@ -63,206 +61,6 @@ impl AutoFill {
 
 // Async API methods are now on Client directly via wsh/async.rs
 
-// Tests that work with both sync and async features
 #[cfg(test)]
-mod common_tests {
-    use super::*;
-    use crate::messages::ResponseMessage;
-
-    #[test]
-    fn test_autofill_is_specified() {
-        assert!(!AutoFill::default().is_specified());
-
-        assert!(AutoFill {
-            competitors: true,
-            portfolio: false,
-            watchlist: false,
-        }
-        .is_specified());
-
-        assert!(AutoFill {
-            competitors: false,
-            portfolio: true,
-            watchlist: false,
-        }
-        .is_specified());
-
-        assert!(AutoFill {
-            competitors: false,
-            portfolio: false,
-            watchlist: true,
-        }
-        .is_specified());
-    }
-
-    #[test]
-    fn test_autofill_combinations() {
-        // Test all possible combinations
-        let combinations = vec![
-            (false, false, false, false),
-            (true, false, false, true),
-            (false, true, false, true),
-            (false, false, true, true),
-            (true, true, false, true),
-            (true, false, true, true),
-            (false, true, true, true),
-            (true, true, true, true),
-        ];
-
-        for (competitors, portfolio, watchlist, expected) in combinations {
-            let autofill = AutoFill {
-                competitors,
-                portfolio,
-                watchlist,
-            };
-            assert_eq!(
-                autofill.is_specified(),
-                expected,
-                "Failed for combination: competitors={}, portfolio={}, watchlist={}",
-                competitors,
-                portfolio,
-                watchlist
-            );
-        }
-    }
-
-    #[test]
-    fn test_decode_wsh_metadata() {
-        use super::decoders::decode_wsh_metadata;
-
-        let message = ResponseMessage::from("104\09000\0{\"test\":\"data\"}\0");
-        let result = decode_wsh_metadata(message);
-
-        assert!(result.is_ok(), "failed to decode wsh metadata: {}", result.err().unwrap());
-        assert_eq!(result.unwrap().data_json, "{\"test\":\"data\"}");
-    }
-
-    #[test]
-    fn test_decode_wsh_event_data() {
-        use super::decoders::decode_wsh_event_data;
-
-        let message = ResponseMessage::from("105\09000\0{\"test\":\"data\"}\0");
-        let result = decode_wsh_event_data(message);
-
-        assert!(result.is_ok(), "failed to decode wsh event data: {}", result.err().unwrap());
-        assert_eq!(result.unwrap().data_json, "{\"test\":\"data\"}");
-    }
-
-    #[test]
-    fn test_encode_request_wsh_metadata() {
-        use super::encoders::encode_request_wsh_metadata;
-        use crate::common::test_utils::helpers::assert_proto_msg_id;
-        use crate::messages::OutgoingMessages;
-
-        let result = encode_request_wsh_metadata(9000);
-        assert!(result.is_ok());
-        assert_proto_msg_id(&result.unwrap(), OutgoingMessages::RequestWshMetaData);
-    }
-
-    #[test]
-    fn test_encode_cancel_wsh_metadata() {
-        use super::encoders::encode_cancel_wsh_metadata;
-        use crate::common::test_utils::helpers::assert_proto_msg_id;
-        use crate::messages::OutgoingMessages;
-
-        let result = encode_cancel_wsh_metadata(9000);
-        assert!(result.is_ok());
-        assert_proto_msg_id(&result.unwrap(), OutgoingMessages::CancelWshMetaData);
-    }
-
-    #[test]
-    fn test_encode_request_wsh_event_data() {
-        use super::encoders::encode_request_wsh_event_data;
-        use crate::common::test_utils::helpers::assert_proto_msg_id;
-        use crate::messages::OutgoingMessages;
-        use time::macros::date;
-
-        // Test with minimal params
-        let result = encode_request_wsh_event_data(9000, Some(12345), None, None, None, None, None);
-        assert!(result.is_ok());
-        assert_proto_msg_id(&result.unwrap(), OutgoingMessages::RequestWshEventData);
-
-        // Test with all params
-        let result = encode_request_wsh_event_data(
-            9000,
-            Some(12345),
-            Some("filter"),
-            Some(date!(2024 - 01 - 01)),
-            Some(date!(2024 - 12 - 31)),
-            Some(100),
-            Some(AutoFill {
-                competitors: true,
-                portfolio: false,
-                watchlist: true,
-            }),
-        );
-        assert!(result.is_ok());
-        assert_proto_msg_id(&result.unwrap(), OutgoingMessages::RequestWshEventData);
-    }
-
-    #[test]
-    fn test_encode_cancel_wsh_event_data() {
-        use super::encoders::encode_cancel_wsh_event_data;
-        use crate::common::test_utils::helpers::assert_proto_msg_id;
-        use crate::messages::OutgoingMessages;
-
-        let result = encode_cancel_wsh_event_data(9000);
-        assert!(result.is_ok());
-        assert_proto_msg_id(&result.unwrap(), OutgoingMessages::CancelWshEventData);
-    }
-
-    #[test]
-    fn test_decode_wsh_metadata_empty_json() {
-        use super::decoders::decode_wsh_metadata;
-
-        let message = ResponseMessage::from("104\09000\0\0");
-        let result = decode_wsh_metadata(message);
-
-        assert!(result.is_ok(), "failed to decode empty wsh metadata: {}", result.err().unwrap());
-        assert_eq!(result.unwrap().data_json, "");
-    }
-
-    #[test]
-    fn test_decode_wsh_event_data_empty_json() {
-        use super::decoders::decode_wsh_event_data;
-
-        let message = ResponseMessage::from("105\09000\0\0");
-        let result = decode_wsh_event_data(message);
-
-        assert!(result.is_ok(), "failed to decode empty wsh event data: {}", result.err().unwrap());
-        assert_eq!(result.unwrap().data_json, "");
-    }
-
-    #[test]
-    fn test_decode_wsh_metadata_with_special_chars() {
-        use super::decoders::decode_wsh_metadata;
-
-        let message = ResponseMessage::from("104\09000\0{\"data\":\"test\\nwith\\tspecial\\rchars\"}\0");
-        let result = decode_wsh_metadata(message);
-
-        assert!(result.is_ok());
-        assert_eq!(result.unwrap().data_json, "{\"data\":\"test\\nwith\\tspecial\\rchars\"}");
-    }
-
-    #[test]
-    fn test_encode_request_wsh_event_data_edge_cases() {
-        use super::encoders::encode_request_wsh_event_data;
-        use crate::common::test_utils::helpers::assert_proto_msg_id;
-        use crate::messages::OutgoingMessages;
-
-        // Test with empty filter string
-        let result = encode_request_wsh_event_data(9000, None, Some(""), None, None, None, None);
-        assert!(result.is_ok());
-        assert_proto_msg_id(&result.unwrap(), OutgoingMessages::RequestWshEventData);
-
-        // Test with special characters in filter
-        let result = encode_request_wsh_event_data(9001, None, Some("filter=\"test\" AND type='earnings'"), None, None, None, None);
-        assert!(result.is_ok());
-        assert_proto_msg_id(&result.unwrap(), OutgoingMessages::RequestWshEventData);
-
-        // Test with negative limit (should still encode)
-        let result = encode_request_wsh_event_data(9002, Some(12345), None, None, None, Some(-10), None);
-        assert!(result.is_ok());
-        assert_proto_msg_id(&result.unwrap(), OutgoingMessages::RequestWshEventData);
-    }
-}
+#[path = "common_tests.rs"]
+mod common_tests;
