+++ todos/test-builders.md
@@ -9,7 +9,7 @@
 | #497 / #498 | Orders (request + simple response builders, sync/async test migration, decoder integration tests, /simplify cleanup) | Merged |
 | #499 | Contracts (request builders, sync/async test migration, encoder self-loop drops, /simplify cleanup) | Merged |
 | #500 | Market data (request builders, sync/async test migration, encoder self-loop drops, /simplify cleanup) | Merged |
-| PR 6 | News, scanner, WSH | Pending |
+| PR 6 | News, scanner, WSH (request builders, sync/async test migration, sibling test files, encoder self-loop drops, dropped unused `MessageBus::request_messages` trait method) | Open |
 
 Foundations that grew beyond the original plan (added during PR 1/PR 2):
 - `ResponseProtoEncoder` trait — symmetric to `RequestEncoder` for the proto path on response builders. Implementors define `Proto` + `to_proto`; trait provides `encode_proto`.
@@ -146,7 +146,7 @@ Per user direction: foundation lands first as a separate PR, then domain-by-doma
 3. ✅ **PR 3 — Orders** ([#497](https://github.com/wboayue/rust-ibapi/pull/497) + [#498](https://github.com/wboayue/rust-ibapi/pull/498), merged). Request builders for place/cancel/(all/auto)open/completed/executions/global_cancel/next_valid_order_id/exercise_options. Simple-response builders for OrderStatus, CommissionReport, ExecutionData, OpenOrderEnd, ExecutionDataEnd, CompletedOrdersEnd. Migrated sync/async tests to `assert_request<B>` body verification. 3 builder→production-decoder integration tests added in `orders/common/decoders/tests.rs`. 11 inline self-loop tests in `orders/common/encoders.rs` dropped per PR 2 lessons. OpenOrder/CompletedOrder builders deferred — those have ~100+ fields and current tests work fine with the existing inline literals. PR 3 also un-gated `MessageBusStub::with_responses` from `#[cfg(feature = "sync")]` so the constructor is available to both sync and async test modules. /simplify pass landed in #498.
 4. ✅ **PR 4 — Contracts** ([#499](https://github.com/wboayue/rust-ibapi/pull/499), merged). Request builders for contract_data, matching_symbols, market_rule, calculate_option_price, calculate_implied_volatility, cancel_contract_data, option_chain. Migrated sync/async tests to `assert_request<B>`. Dropped encoder self-loop tests. Response builders deferred (current inline literals already exercise the production decoders end-to-end).
 5. ✅ **PR 5 — Market data** ([#500](https://github.com/wboayue/rust-ibapi/pull/500), merged). Request builders for historical (head_timestamp, historical_data, historical_ticks, histogram_data) and realtime (realtime_bars, tick_by_tick, market_depth, market_depth_exchanges, market_data). Migrated sync/async tests in both submodules to `assert_request<B>` body verification. Dropped 17 inline encoder self-loop tests (kept `test_encode_interval` — exercises the `OffsetDateTime: ToField` impl). Cancel builders dropped during /simplify pass (no test consumer); revisit when a `test_cancel_*` body assertion lands. Response builders deferred — market-data responses are text wire only and the existing inline literals exercise the production decoders.
-6. **PR 6 — News, scanner, WSH.**
+6. ⏳ **PR 6 — News, scanner, WSH** (open). Request builders: news (providers, bulletins, cancel-bulletins, historical, article — `contract_news`/`broad_tape_news` reuse `MarketDataRequestBuilder`); scanner (parameters, subscription, cancel); wsh (metadata, event-data, cancels). Sync/async tests migrated to `assert_request<B>` body verification. Inline `mod tests` blocks moved to sibling `_tests.rs` files (`news/{sync,async}_tests.rs`, `scanner/{sync,async}_tests.rs`, `wsh/{sync,async}_tests.rs`, `wsh/common_tests.rs`). Self-loop encoder tests dropped from `{news,scanner,wsh}/common/encoders.rs` and from `wsh/mod.rs::common_tests`. Removed dead `MessageBus::request_messages` and `AsyncMessageBus::request_messages` trait methods (and their `MessageBusStub` impls) — the only consumers were the migrated sync tests, which now use the inherent `MessageBusStub::request_messages` via `assert_request<B>(&message_bus, …)`.
 
 Each domain PR is small enough to review on its own and surfaces builder API gaps incrementally. If PR 2's adoption reveals an awkward API, fix in PR 2 before propagating.
